Added text files for tests. Added classes for maze generation and solving.
