+++ src/main/java/edu/project2/MazeGenerator.java
@@ -2,58 +2,17 @@ package edu.project2;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Random;
 
-public class MazeGenerator {
+public abstract class MazeGenerator {
 
-    private static final Random RANDOM = new Random();
+    protected static final Random RANDOM = new Random();
 
-    private static final ArrayList<Point> POINTS = new ArrayList<>(
+    protected static final ArrayList<Point> POINTS = new ArrayList<>(
         List.of(new Point(2, 0), new Point(-2, 0), new Point(0, 2), new Point(0, -2)));
 
-    public Maze createMazeRB(int horizontalCellsAmount, int verticalCellsAmount, int seedRandom) {
-        if (seedRandom != -1) {
-            RANDOM.setSeed(seedRandom);
-        }
-
-        Maze maze = createMaze(horizontalCellsAmount, verticalCellsAmount, Maze.WALL);
-
-        createMazeRecursiveBacktracking(maze, maze.getStartPoint());
-
-        createEdgePoints(maze);
-
-        return maze;
-    }
-
-    private void createMazeRecursiveBacktracking(Maze maze, Point point) {
-        Point newPoint;
-        Point move;
-        ArrayList<Point> points = new ArrayList<>(POINTS);
-
-        Collections.shuffle(points, RANDOM);
-
-        for (int i = 0, size = points.size(); i < size; i++) {
-            newPoint = point.getCopy();
-
-            move = points.get(i);
-
-            newPoint.add(move);
-
-            if (maze.isInMaze(newPoint)
-                && maze.getCellInMaze(newPoint) != Maze.EMPTY_BLOCK) {
-
-                removeBlocks(maze, newPoint.getX() - move.getX() / 2, newPoint.getY() - move.getY() / 2);
-                removeBlocks(maze, newPoint.getX(), newPoint.getY());
-                createMazeRecursiveBacktracking(maze, newPoint);
-            }
-        }
-    }
-
-    private void removeBlocks(Maze maze, int x, int y) {
+    protected void removeBlocks(Maze maze, int x, int y) {
         maze.setCellInMaze(x, y, Maze.EMPTY_BLOCK);
     }
 
@@ -78,14 +37,14 @@ public class MazeGenerator {
         return new Point(x, y);
     }
 
-    private void createEdgePoints(Maze maze) {
+    protected void createEdgePoints(Maze maze) {
         maze.setStartAndEndPoints(
             createBlankPoint(maze, maze.getStartPoint()),
             createBlankPoint(maze, maze.getEndPoint())
         );
     }
 
-    private Point createBlankPoint(Maze maze, Point point) {
+    protected Point createBlankPoint(Maze maze, Point point) {
         Point edgePoint;
 
         if (point.getX() > 0 && point.getX() < maze.getWidth() - 1) {
@@ -113,7 +72,7 @@ public class MazeGenerator {
         return edgePoint;
     }
 
-    private Maze createMaze(int horizontalCellsAmount, int verticalCellsAmount, char symbol) {
+    protected Maze createMazePreparation(int horizontalCellsAmount, int verticalCellsAmount, char symbol) {
         Maze maze = new Maze(horizontalCellsAmount, verticalCellsAmount);
 
         fillMaze(maze, symbol);
@@ -130,135 +89,5 @@ public class MazeGenerator {
         return maze;
     }
 
-    public Maze createMazeEA(int horizontalCellsAmount, int verticalCellsAmount, int seedRandom) {
-        if (seedRandom != -1) {
-            RANDOM.setSeed(seedRandom);
-        }
-
-        Maze maze = createMaze(horizontalCellsAmount, verticalCellsAmount, Maze.EMPTY_BLOCK);
-
-        createMazeEllerAlgorithm(maze);
-
-        createEdgePoints(maze);
-
-        return maze;
-    }
-
-    private void createMazeEllerAlgorithm(Maze maze) {
-
-        int[] mazeRow = new int[(maze.getWidth() - 1) / 2];
-        int rowsAmount = (maze.getHeight() - 1) / 2;
-        int setNumber = 1;
-
-        for (int i = 0; i < rowsAmount; i++) {
-            // step 1
-            setNumber = setNumbersInRow(setNumber, mazeRow);
-
-            int y = i * 2 + 1;
-
-            // step 2 - right walls
-            addRightWalls(y, mazeRow, maze);
-
-            // step 3 - bottom walls
-            addBottomWalls(y, mazeRow, maze);
-
-            // step 4
-            finishOrPrepareForNextRow(i, y, rowsAmount, mazeRow, maze);
-        }
-        completeMaze(maze);
-    }
-
-    private int setNumbersInRow(int setNumber, int[] mazeRow) {
-        int tmpSetNumber = setNumber;
-
-        for (int j = 0; j < mazeRow.length; j++) {
-            if (mazeRow[j] == 0) {
-                mazeRow[j] = tmpSetNumber++;
-            }
-        }
-
-        return tmpSetNumber;
-    }
-
-    private void addRightWalls(int y, int[] mazeRow, Maze maze) {
-        for (int j = 0; j < mazeRow.length - 1; j++) {
-            if ((mazeRow[j + 1] == mazeRow[j]) || RANDOM.nextBoolean()) {
-                maze.setCellInMaze(j * 2 + 2, y, Maze.WALL);
-            } else {
-                // combine sets
-                int changeSet = mazeRow[j + 1];
-
-                for (int n = 0; n < mazeRow.length; n++) {
-                    if (mazeRow[n] == changeSet) {
-                        mazeRow[n] = mazeRow[j];
-                    }
-                }
-            }
-        }
-    }
-
-    private void addBottomWalls(int y, int[] mazeRow, Maze maze) {
-        Map<Integer, Integer> map = new HashMap<>();
-        for (int j = 0; j < mazeRow.length; j++) {
-            if (map.containsKey(mazeRow[j])) {
-                map.put(mazeRow[j], map.get(mazeRow[j]) + 1);
-            } else {
-                map.put(mazeRow[j], 1);
-            }
-        }
-
-        for (int j = 0; j < mazeRow.length - 1; j++) {
-            if (map.get(mazeRow[j]) > 1) {
-                if (RANDOM.nextBoolean()) {
-                    maze.setCellInMaze(j * 2 + 1, y + 1, Maze.WALL);
-                    map.put(mazeRow[j], map.get(mazeRow[j]) - 1);
-                }
-            }
-        }
-    }
-
-    private void finishOrPrepareForNextRow(int i, int y, int rowsAmount, int[] mazeRow, Maze maze) {
-        if (i == rowsAmount - 1) {
-            for (int j = 0; j < mazeRow.length - 1; j++) {
-                if (mazeRow[j + 1] != mazeRow[j]) {
-                    maze.setCellInMaze(j * 2 + 2, y, Maze.EMPTY_BLOCK);
-
-                    // full merging of two sets
-                    int changeSet = mazeRow[j + 1];
-                    for (int n = 0; n < mazeRow.length; n++) {
-                        if (mazeRow[n] == changeSet) {
-                            mazeRow[n] = mazeRow[j];
-                        }
-                    }
-                }
-            }
-        } else {
-            for (int j = 0; j < mazeRow.length; j++) {
-                if (maze.getCellInMaze(j * 2 + 1, y + 1) == Maze.WALL) {
-                    mazeRow[j] = 0;
-                }
-            }
-        }
-    }
-
-    private void completeMaze(Maze maze) {
-        int height = maze.getHeight();
-        int width = maze.getWidth();
-
-        for (int x = 0; x < width; x++) {
-            maze.setCellInMaze(x, 0, Maze.WALL);
-            maze.setCellInMaze(x, height - 1, Maze.WALL);
-        }
-
-        for (int y = 1; y < height - 1; y++) {
-            maze.setCellInMaze(0, y, Maze.WALL);
-            maze.setCellInMaze(width - 1, y, Maze.WALL);
-        }
-
-        for (int y = 2; y < height - 1; y += 2) {
-            for (int x = 2; x < width - 1; x += 2) {
-                maze.setCellInMaze(x, y, Maze.WALL);
-            }
-        }
-    }
+    public abstract Maze createMaze(int horizontalCellsAmount, int verticalCellsAmount, int seedRandom);
 }
