+++ src/test/java/edu/project2/MazeSolverTest.java
@@ -7,48 +7,49 @@ import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class MazeSolverTest {
-
-    private MazeSolver mazeSolver;
     private LinkedList<Point> wayToExit;
 
     @BeforeEach
     void setUp() {
-        mazeSolver = new MazeSolver();
         wayToExit = new LinkedList<>();
     }
 
     @Test
     void findBreadthFirstWay_WhenGivenSolvableMaze_SolvesMaze() {
-        Maze maze = Maze.getInstanceFromFile(new File("src/main/resources/project2/maze-2.txt"));
+        MazeSolver mazeSolverBF = new MazeSolverBF();
+        Maze maze = Maze.getInstanceFromFile(new File("src/test/resources/project2/maze-2.txt"));
 
-        mazeSolver.findBFWay(maze, maze.getStartPoint(), maze.getEndPoint(), wayToExit);
+        mazeSolverBF.findWay(maze, maze.getStartPoint(), maze.getEndPoint(), wayToExit);
 
         assertThat(wayToExit).isNotEmpty();
     }
 
     @Test
     void findBreadthFirstWay_WhenGivenUnsolvableMaze_DoesNotSolveMaze() {
-        Maze maze = Maze.getInstanceFromFile(new File("src/main/resources/project2/maze-3.txt"));
+        MazeSolver mazeSolverBF = new MazeSolverBF();
+        Maze maze = Maze.getInstanceFromFile(new File("src/test/resources/project2/maze-3.txt"));
 
-        mazeSolver.findBFWay(maze, maze.getStartPoint(), maze.getEndPoint(), wayToExit);
+        mazeSolverBF.findWay(maze, maze.getStartPoint(), maze.getEndPoint(), wayToExit);
 
         assertThat(wayToExit).isEmpty();
     }
 
     @Test
     void findDepthFirstWay_WhenGivenSolvableMaze_SolvesMaze() {
-        Maze maze = Maze.getInstanceFromFile(new File("src/main/resources/project2/maze-2.txt"));
+        MazeSolver mazeSolverDF = new MazeSolverDF();
+        Maze maze = Maze.getInstanceFromFile(new File("src/test/resources/project2/maze-2.txt"));
 
-        mazeSolver.findDFWay(maze, maze.getStartPoint(), maze.getEndPoint(), wayToExit);
+        mazeSolverDF.findWay(maze, maze.getStartPoint(), maze.getEndPoint(), wayToExit);
 
         assertThat(wayToExit).isNotEmpty();
     }
 
     @Test
     void findDepthFirstWay_WhenGivenUnsolvableMaze_DoesNotSolveMaze() {
-        Maze maze = Maze.getInstanceFromFile(new File("src/main/resources/project2/maze-3.txt"));
+        MazeSolver mazeSolverDF = new MazeSolverDF();
+        Maze maze = Maze.getInstanceFromFile(new File("src/test/resources/project2/maze-3.txt"));
 
-        mazeSolver.findDFWay(maze, maze.getStartPoint(), maze.getEndPoint(), wayToExit);
+        mazeSolverDF.findWay(maze, maze.getStartPoint(), maze.getEndPoint(), wayToExit);
 
         assertThat(wayToExit).isEmpty();
     }
