+++ src/test/resources/project2/maze-1.txt
@@ -1,0 +1,21 @@
+█████████████████████████████████████████████████████████████
+█           █           █ █   █             █   █       █   █
+█ ███████ ███ █████ ███ █ █ █ █ ███ ███████ █ █ █████ █ █ █ █
+█ █       █   █     █ █   █ █ █ █ █       █   █ █   █ █   █ █
+█ █ ███████ ███ █████ ███ █ █ █ █ ███████ █████ █ █ █ ███████
+█ █ █       █   █     █   █ █   █     █   █ █     █   █     █
+█ ███ ███████ ███ ███ █ ███ █████ █ ███ ███ █ █████████ ███ █
+█       █   █ █ █   █   █ █ █   █ █ █   █       █   █   █ █ █
+█ ███████ █ █ █ ███ █████ █ █ █ ███ █ █ █████████ █ █ ███ █ █
+█   █ █   █   █ █   █     █   █     █ █ █         █   █   █ █
+███ █ █ ███████ █ ███ ███ █████████ █ ███ ███████████ █ ███ █
+█ █ █ █ █   █     █     █         █ █     █     █   █   █   █
+█ █ █ █ █ █ █ █████████ ███████████ ███████ ███ █ █ █████ ███
+█   █ █ █ █ █ █     █   █     █     █     █ █   █ █     █   █
+█ ███ █ █ ███ █ ███ █ █ █ ███ █ ███████ █ █ █████ █████ █ █ █
+█ █   █ █     █ █ █   █ █   █ █       █ █   █     █   █ █ █ █
+█ █ █ █ █████ █ █ █ ███████ █ ███████ █ █████ ███████ █ ███ █
+█ █ █ █     █   █   █     █ █   █   █ █     █   █     █   █ █
+█ █ █ █████ █████████ ███ █ █ ███ █ █ █████ ███ █ █ █████ █ █
+█   █     █             █   █     █         █   █ █         █
+███████████████████████████████████████████A███B█████████████
