+++ src/test/resources/project2/maze-3.txt
@@ -1,0 +1,11 @@
+███████████
+█         █
+█ █ █ █ █ █
+█         █
+█ █ █ ███ █
+█     █B█ █
+█ █ █ ███ █
+█  A      █
+█ █ █ █ █ █
+█         █
+███████████
