+++ src/test/resources/project2/maze-invalid-format.txt
@@ -1,0 +1,11 @@
+███████████
+█         █
+█ █ █ █ █ █
+█         █
+█ █ █ ███ █
+█      B  █
+█ █ █ ███ █
+█  A      █
+█ █ █ █ █ █
+█      A  █
+███████████
