+++ src/test/resources/project2/maze-invalid-size.txt
@@ -1,0 +1,19 @@
+████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████
+█                                                                                                                                                                                         ██████████████████████████████████████████████████████████████████████████████████
+█                                                                                                                                                                                         ██████████████████████████████████████████████████████████████████████████████████
+█                                                                                                                                                                                         ██████████████████████████████████████████████████████████████████████████████████
+█                                                                                                                                                                                         ██████████████████████████████████████████████████████████████████████████████████
+█                                                                                                                                                                                         ██████████████████████████████████████████████████████████████████████████████████
+█                                                                                                                                                                                         ██████████████████████████████████████████████████████████████████████████████████
+█                                                                                                                                                                                         ██████████████████████████████████████████████████████████████████████████████████
+█                                                                                                                                                                                         ██████████████████████████████████████████████████████████████████████████████████
+█                                                                                                                                                                                         ██████████████████████████████████████████████████████████████████████████████████
+█                                                                                                                                                                                         ██████████████████████████████████████████████████████████████████████████████████
+█                                                                                                                                                                                         ██████████████████████████████████████████████████████████████████████████████████
+█                                                                                                                                                                                         ██████████████████████████████████████████████████████████████████████████████████
+█                                                                                                                                                                                         ██████████████████████████████████████████████████████████████████████████████████
+█                                                                                                                                                                                         ██████████████████████████████████████████████████████████████████████████████████
+█                                                                                                                                                                                         ██████████████████████████████████████████████████████████████████████████████████
+█                                                                                                                                                                                         ██████████████████████████████████████████████████████████████████████████████████
+█                                                                                                                                                                                         ██████████████████████████████████████████████████████████████████████████████████
+████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████
